refactored the player
