+++ .idea/workspace.xml
@@ -4,30 +4,18 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="d7b308d3-29c1-431e-9c6d-76e5d7a2b191" name="Changes" comment="made player icons teleport tiles one by one">
-      <change afterPath="$PROJECT_DIR$/src/main/entity/players/StockModifier.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/entity/players/applyAfterEffects.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/entity/players/rentModifier.java" afterDir="false" />
+    <list default="true" id="d7b308d3-29c1-431e-9c6d-76e5d7a2b191" name="Changes" comment="refactored the player">
+      <change afterPath="$PROJECT_DIR$/src/main/use_case/Stocks/BuyStock.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/main/use_case/Stocks/SellStock.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/Constants/Constants.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/Constants/Constants.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/entity/players/DefaultPlayer.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/DefaultPlayer.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/players/Player.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/Player.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/entity/players/PoorMan.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/PoorMan.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/entity/players/clerk.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/clerk.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/entity/players/collegeStudent.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/collegeStudent.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/entity/players/inheritor.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/inheritor.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/entity/players/landlord.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/landlord.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/entity/tiles/PropertyTile.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/tiles/PropertyTile.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/entity/tiles/StockMarketTile.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/tiles/StockMarketTile.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/interface_adapter/StockMarket/StockMarketViewModel.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/interface_adapter/StockMarket/StockMarketViewModel.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/use_case/Player.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/players/Player.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/use_case/Tile.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/use_case/Tile.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/view/BoardRenderer.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/BoardRenderer.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/view/BoardView.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/BoardView.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/view/BuyPropertyPopup.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/BuyPropertyPopup.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/view/EndScreen.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/EndScreen.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/view/GameBoard.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/GameBoard.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/view/PlayerMovementAnimator.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/PlayerMovementAnimator.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/view/PlayerStatsView.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/PlayerStatsView.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/view/StockMarketView.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/StockMarketView.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/view/StockView.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/view/StockView.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -38,6 +26,7 @@
     <option name="RECENT_TEMPLATES">
       <list>
         <option value="Interface" />
+        <option value="Class" />
       </list>
     </option>
   </component>
@@ -165,7 +154,7 @@
       <workItem from="1752620400736" duration="12322000" />
       <workItem from="1753307514943" duration="5000" />
       <workItem from="1753308904677" duration="5497000" />
-      <workItem from="1753794843328" duration="6567000" />
+      <workItem from="1753794843328" duration="11801000" />
     </task>
     <task id="LOCAL-00001" summary="added &quot;teleportation movement of players&quot;">
       <option name="closed" value="true" />
@@ -199,7 +188,15 @@
       <option name="project" value="LOCAL" />
       <updated>1753397991589</updated>
     </task>
-    <option name="localTasksCounter" value="5" />
+    <task id="LOCAL-00005" summary="refactored the player">
+      <option name="closed" value="true" />
+      <created>1753825699937</created>
+      <option name="number" value="00005" />
+      <option name="presentableId" value="LOCAL-00005" />
+      <option name="project" value="LOCAL" />
+      <updated>1753825699937</updated>
+    </task>
+    <option name="localTasksCounter" value="6" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -221,6 +218,7 @@
     <MESSAGE value="made player icons teleport tiles one by one" />
     <MESSAGE value="minor bug fixes" />
     <MESSAGE value="Player changes colour" />
-    <option name="LAST_COMMIT_MESSAGE" value="Player changes colour" />
+    <MESSAGE value="refactored the player" />
+    <option name="LAST_COMMIT_MESSAGE" value="refactored the player" />
   </component>
 </project>
