+++ src/main/Constants/Constants.java
@@ -45,4 +45,14 @@ public class Constants {
     // PlayerStatesView
     public static final int DIMENSION_DE_LA_PSV_H = 200;
     public static final int DIMENSION_DE_LA_PSV_W = 200;
+
+    // Character Constants
+    public static final int CLERK_INIT_MONEY = 1200;
+    public static final int STUDENT_INIT_MONEY = 1000;
+    public static final int DEFAULT_INIT_MONEY = 1200;
+    public static final int INHERITOR_INIT_MONEY = 1800;
+    public static final int LANDLORD_INIT_MONEY = 800;
+    public static final int POORMAN_INIT_MONEY = 200;
+
+    public static final int CLERK_ADD_MONEY = 50;
 }
