+++ src/main/entity/players/DefaultPlayer.java
@@ -4,8 +4,9 @@ import main.entity.Stock;
 
 import java.awt.*;
 
+import static main.Constants.Constants.DEFAULT_INIT_MONEY;
+
 public class DefaultPlayer extends Player {
-    private static final int DEFAULT_INIT_MONEY = 1200;
     private String name;
 
     public DefaultPlayer(String name, Color color) {
@@ -23,24 +24,5 @@ public class DefaultPlayer extends Player {
 
     public int getPosition() {
         return this.position;
-    }
-
-    @Override
-    public void buyStock(Stock stock, int quantity) {
-        double totalCost = stock.getCurrentPrice() * quantity;
-
-        if (this.money >= totalCost) {
-            this.deductMoney((float) totalCost);
-            stocks.put(stock, stocks.getOrDefault(stock, 0) + quantity);
-        }
-    }
-
-    @Override
-    public void sellStock(Stock stock, int quantity) {
-        if (stocks.getOrDefault(stock, 0) >= quantity) {
-            double totalSale = stock.getCurrentPrice() * quantity;
-            this.addMoney((float) totalSale);
-            stocks.put(stock, stocks.get(stock) - quantity);
-        }
     }
 }
